add:open red packet automatically

+++ app/src/main/java/com/newyear/redbull/model/AppViewModel.kt
@@ -29,7 +29,7 @@ class AppViewModel (
         updateBasicFunctionalityCache()
     }
 
-    fun updateDelaySeconds(value: Float) {
+    fun updateDelaySeconds(value: Int) {
         _basicFunctionalityState.update { currentState ->
             currentState.copy(
                 delaySeconds = value
